provide Keydata.get_idx_version
In particular, the idx is unique at any one instant, and "small".
There are possible use cases where this is a useful property and will
save the caller the trouble of maintaining a separate set of ids.

It seems churlish to hide this, particularly given as it's even
printed by the Debug impl.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>